+++ inst/templates/HEADER.R
@@ -3,5 +3,7 @@
 #+ reprex-setup, include = FALSE
 knitr::opts_chunk$set(collapse = TRUE, comment = '#>', error = TRUE)
 knitr::opts_knit$set(upload.fun = knitr::imgur_upload)
+{{{user_opts_chunk}}}
+{{{user_opts_knit}}}
 
 #+ reprex-body